fix(dmx): add full RGBW support and fix 512-channel limitation
- Add RGBW spot type support (4 channels: Red, Green, Blue, White)
- Fix critical bug: change start_channel from uint8_t to uint16_t
  * Previous limitation: only channels 0-255 accessible
  * Now supports full DMX universe (0-512 channels)
- Fix DMX sending thread to handle RGBW spots correctly
  * Thread now detects spot type (RGB vs RGBW)
  * Sends 4 channels for RGBW instead of only 3
- Implement White channel calculation using min(R,G,B) algorithm
  * Preserves color hue while adding brightness
- Configure for 90 RGBW LEDs (5 bars × 18 LEDs = 360 channels)
- Add RGBW initialization in dmx_generate_channel_mapping()
- Add RGBW validation in dmx_init_configuration()

This enables proper control of Pixel Bar 18Q4 Tour fixtures
on channels 1-72, 73-144, 145-216, 217-288, and 289-360.

+++ src/communication/dmx/dmx.h
@@ -43,7 +43,7 @@ typedef union {
 // Main spot structure with type and start channel
 typedef struct {
     DMXSpotType type;
-    uint8_t start_channel;  // DMX start channel for this spot
+    uint16_t start_channel;  // DMX start channel for this spot (supports 0-65535)
     DMXSpotData data;
 } DMXSpot;
